+++ src/main/common/form/models/claim/details/claimDetails.ts
@@ -1,9 +1,12 @@
 import {Reason} from './reason';
+import {HelpWithFees} from './helpWithFees';
 
 export class ClaimDetails {
   reason?: Reason;
+  helpWithFees?: HelpWithFees;
 
-  constructor(reason?: Reason) {
+  constructor(reason?: Reason, helpWithFees?: HelpWithFees) {
     this.reason = reason;
+    this.helpWithFees = helpWithFees;
   }
 }
