+++ src/main/views/features/claim/details/help-with-fees.njk
@@ -1,11 +1,11 @@
 {% extends "claim-details-tpl.njk" %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 {% from "govuk/components/input/macro.njk" import govukInput %}
-{% from "../../macro/contact-us-for-help.njk" import contactUsForHelp %}
-{% from "../../macro/csrf.njk" import csrfProtection %}
-{% from "../../macro/yesNoGenericForm.njk" import yesNoRadioButton %}
-{% from "../../macro/page-title.njk" import setPageTitle %}
-{% from "../../macro/error-summary.njk" import errorSummary %}
+{% from "../../../macro/contact-us-for-help.njk" import contactUsForHelp %}
+{% from "../../../macro/csrf.njk" import csrfProtection %}
+{% from "../../../macro/yesNoGenericForm.njk" import yesNoRadioButton %}
+{% from "../../../macro/page-title.njk" import setPageTitle %}
+{% from "../../../macro/error-summary.njk" import errorSummary %}
 
 {% block pageTitle %}
   {{ setPageTitle('PAGES.HELP_WITH_FEES.PAGE_TITLE') }}
